Modularize UI code
