+++ lib/src/app.dart
@@ -6,7 +6,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'sample_feature/sample_item_details_view.dart';
 import 'sample_feature/sample_item_list_view.dart';
 import 'settings/settings_controller.dart';
-import 'settings/settings_view.dart';
+import 'settings/screens/settings_view/settings_view.dart';
 
 /// The Widget that configures your application.
 class MyApp extends StatelessWidget {
