+++ lib/src/sample_feature/sample_item_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../settings/settings_view.dart';
+import '../settings/screens/settings_view/settings_view.dart';
 import 'sample_item.dart';
 import 'sample_item_details_view.dart';
